#656 Added put(), root_hash() and has_key?() functions to the state_trees module
